Simplify kernel generation.
